Add one-second time-delay between inputs
This ensures that the entire page has loaded before
attempting to locate any elements (WebDriverWait was
not working for some reason). It also fixes a problem
where input for the password text box appears in the
username text box because the computer didn't wait for
the focus change animation to complete before continuing
typing.